requirements.txt
Django==1.10.1
argparse==1.2.1
djangorestframework==3.4.1
wsgiref==0.1.2
